merge 1.0.51 into master.

+++ includes/features/functions.php
@@ -381,7 +381,10 @@ function ampforwp_generate_meta_desc($json=""){
         // Rank Math SEO #2701
         if ( defined( 'RANK_MATH_FILE' ) && 'rank_math' == ampforwp_get_setting('ampforwp-seo-selection') ) {
             $rank_math_desc = RankMath\Post::get_meta( 'description', $post_id );
-            $desc           = $rank_math_desc ? $rank_math_desc : $desc;
+            if (ampforwp_is_home()) {
+                $rank_math_desc = $desc;
+            }
+            $desc = $rank_math_desc ? $rank_math_desc : $desc;
         }
         //Bridge Qode SEO Compatibility #2538 
         if ( function_exists('qode_header_meta') && 'bridge' == ampforwp_get_setting('ampforwp-seo-selection')){
@@ -1369,7 +1372,7 @@ function ampforwp_category_base_remove_notice(){
     if(true == ampforwp_get_setting('ampforwp-category-base-removel-link')){
         return;
     }
-    if(class_exists('WPSEO_Options') && WPSEO_Options::get( 'stripcategorybase' ) == true){
+    if(class_exists('WPSEO_Options') && method_exists('WPSEO_Options', 'get') && WPSEO_Options::get( 'stripcategorybase' ) == true ){
         echo sprintf(('<div class="notice notice-error"><p>%s <a href="%s">%s</a>%s<a href="%s">%s</a></p></div>'), esc_html__('We have detected that you removed Category Base from Yoast SEO plugin but not from AMPforWP settings. Please','accelerated-mobile-pages'),
         esc_url('admin.php?page=amp_options&tab=17'),
         esc_html__('Click here','accelerated-mobile-pages'),
